REINICIO: Refactorización y actualización integral del proyecto
Este commit representa un punto de reinicio significativo para el proyecto, donde se han implementado refactorizaciones extensas, mejoras de rendimiento y actualizaciones críticas. Todos los cambios están orientados a establecer una nueva base sólida para el desarrollo futuro

+++ src/main/java/com/almatec/controlpiso/integrapps/dtos/OpProduccionDTO.java
@@ -10,7 +10,7 @@ public class OpProduccionDTO {
 	private String cliente;
 	private String itemRef;
 	private String itemDescripcion;
-	private Integer cant;
+	private Double cant;
 	
 	public OpProduccionDTO() {
 		super();
@@ -69,10 +69,10 @@ public class OpProduccionDTO {
 	public void setItemDescripcion(String itemDescripcion) {
 		this.itemDescripcion = itemDescripcion;
 	}
-	public Integer getCant() {
+	public Double getCant() {
 		return cant;
 	}
-	public void setCant(Integer cant) {
+	public void setCant(Double cant) {
 		this.cant = cant;
 	} 
 	@Override
